refactor(cli): split resolution path mapping tests
